Add simple docker compose configuration
Docker compose was added for easier setup of local environment that
doesn't require developer to create accounts in external services to run
the project. It will be also helpful when we start adding more services
(for example for sending emails). It's also easier to start contributing
to project when you don't have to create external services accounts,
change server versions (and keep them consistent accross engineers).

Project is still using expternal services that should be removed:
 - supabase for database
 - cloudinary for image storage

Signed-off-by: Wojciech Szarański <wojciech.szaranski@gmail.com>